# Speedup xml replacement