--- a/packages/tool_util/requirements.txt
+++ b/packages/tool_util/requirements.txt
@@ -1,2 +1,3 @@
 galaxy-util>=20.1.0.dev0
 galaxy-containers
+lxml