@mvdbeek 2020-11-25
Yes, but ... it passes in the unit tests! Does tool_util not use lxml ?