@mvdbeek 2020-11-25
If I completely disable the tool document cache I'm at 20 seconds for loading main's toolbox with this (vs 12.8 seconds with cache fully populated) ... that's both a lot and not a lot depending on how you look at it. It does make me feel better about making the tool document cache optional or disabling it when there are locking problems on NFS/CIFS or other network shares.