@bgruening 2020-12-05
Looks good to me. In our limited testing, this also works on EU and I would appreciate making the cache optional. We have still huge memory spikes when reloading the toolbox - which I think is due to the new caching system - but not sure.