@mvdbeek 2020-12-07
> We have still huge memory spikes when reloading the toolbox - which I think is due to the new caching system - but not sure.

This seems unlikely to be caused by the cache, nothing is kept in memory there.

> I would appreciate making the cache optional.

You can point the cache to a read-only location, it will not be used then.